Simplify naming - extract logic to service for Pearls
