+++ eternalcombat-api/src/main/java/com/eternalcode/combat/EternalCombatApi.java
@@ -4,7 +4,7 @@ import com.eternalcode.combat.fight.drop.DropKeepInventoryService;
 import com.eternalcode.combat.fight.FightManager;
 import com.eternalcode.combat.fight.drop.DropService;
 import com.eternalcode.combat.fight.effect.FightEffectService;
-import com.eternalcode.combat.fight.pearl.FightPearlService;
+import com.eternalcode.combat.fight.pearl.PearlService;
 import com.eternalcode.combat.region.RegionProvider;
 import com.eternalcode.combat.fight.tagout.FightTagOutService;
 
@@ -14,7 +14,7 @@ public interface EternalCombatApi {
 
     RegionProvider getRegionProvider();
 
-    FightPearlService getFightPearlService();
+    PearlService getFightPearlService();
 
     FightTagOutService getFightTagOutService();
