+++ eternalcombat-api/src/main/java/com/eternalcode/combat/fight/pearl/PearlService.java
@@ -3,11 +3,13 @@ package com.eternalcode.combat.fight.pearl;
 import java.time.Duration;
 import java.util.UUID;
 
-public interface FightPearlService {
+public interface PearlService {
+
+    boolean shouldCancelEvent(UUID playerId);
 
     Duration getRemainingDelay(UUID uuid);
 
     boolean hasDelay(UUID uuid);
 
-    void markDelay(UUID uuid);
+    void handleDelay(UUID uuid);
 }
